// src/app/single-show/single-show.type.ts
export class SingleShowType {
    adult: boolean;
    backdropPath: string;
    belongsToCollection: {};
    budget: number;
    genres: [string];
    homepage: string;
    id: number;
    imdbId: string;
    originalLanguage: string;
    originalTitle: string;
    overview: string;
    popularity: number;
    posterPath: string;
    productionCompanies: [string];
    productionCountries: [string];
    releaseDate: string; // might need to be date?
    revenue: number;
    runtime: number;
    spokenLanguages: [string];
    status: string;
    tagline: string;
    title: string;
    video: boolean;
    voteAverage: number;
    voteCount: number;
}